WIP: Refactor app.
The repo was originally forked from an app develop for Python 3.5 and
Django 1.11.
The goal is to refactor the app for Python 3.11 and Django 4.2.

This commit adds the Django management scripts, as created with

```bash
django-admin startproject setup .
```

This commit also adds the `requirements.txt` file, for tracking the
libraries used in the virtual development environment.