WIP: Refactor. Add initial structure for `pte` application.
The goal is to refactor the content of the `django_periodic_table` app
into a new app named `pte`.
This commit adds the underlying structure of the new `pte` app.

This commit also updates `.gitignore`, so production files are not added
to the development repository.

The `requirements.txt` file is also updated with `python-dotenv`, which
will be used to store applications keys that should not be shown in the
repository.